refactor: 분석 훅을 hooks 디렉토리로 이동 및 임포트 경로 통일
- useInitAnalytics, useTrackPage 훅을 components/analytics에서 hooks/analytics로 이동
- 모든 페이지에서 분석 훅 임포트 경로 업데이트
- 분석 라이브러리 의존성 정리
