+++ apps/web/src/app/onboarding/profile/page.tsx
@@ -9,7 +9,7 @@ import { useProfileUpload } from '../_hooks/useProfileUpload';
 import { useOnboardingContext } from '../_contexts/OnboardingContext';
 import { ROUTES } from '@/constants/routes';
 import { track } from '@/lib/analytics';
-import { useTrackPage } from '@/components/analytics/useTrackPage';
+import { useTrackPage } from '@/hooks/analytics/useTrackPage';
 import * as S from './page.styles';
 
 export default function ProfilePage() {
