+++ apps/web/src/hooks/analytics/useInitAnalytics.ts
